week4_application_layer_database_connection/clarke_csc6302_database_principles_project_mod4/clarke_MRCapplication/requirements.txt
mysql-connector-python==9.5.0
pandas==2.3.3